split filter and page buttons to seperate components
